refactor(ui): create admin sidebar header and footer
